new tests

+++ results/Gemini/APPS/Zero-shot-CoT/code_71.py
@@ -1,4 +1,4 @@
 import math
 
 a = int(input())
-print(int(math.ceil(math.log(a, 2))))
+print(math.ceil(math.log(a, 2)))
